extract base elements from Cruise sign into a toggle sign
